test(topic_entity_tag): Add additional assertions for check on second curation
This commit adds more assertions to the test_topic_entity_tag test suite. Specifically, it checks if the returned values of "validation_value_author", "validation_value_curator" and "validation_value_curation_tools" for the second curator's tag are as expected. This is to further ensure the correct functionality of the /topic_entity_tag/ endpoint.