Pin specific version for types-requests and update requests version
In requirements.dev.txt, types-requests was not initially pinned to a specific version which may cause unexpected behaviors. We've pinned it down to a stable version 2.31.0.6 to maintain consistency in our development environment. Meanwhile, in requirements.txt, we've upgraded the 'requests' library version from 2.28.2 to 2.30.0, while fixing the 'urllib3' version to 1.26.15. This was done to resolve some version conflicts with other packages.