Improve cross-ref search results limit handling and ordering
Made an improvement to the handling of limit results in the cross-reference search within `cross_reference_crud.py`. Now, if matching cross-references count is less than 20, an expanded query is executed, and matching curies are extended with the results. Also, results are ordered by `CrossReferenceModel.curie` to provide a structured display of data.