Fixed conversion from sqlalchemy db obj to dict + updated tests to use new tet post return values
Calling jsonable_encoder on a db obj with loops in the relationships was returning a max recursion error. The exclude parameter of jsonable_encoder only works if the passed object is in dict format. This looks like a bug in the decoder function. Made a first conversion with vars and then called jsonable_encoder with the problematic relationship (validated_by) in the exclude list.

Updated tet web tests to use the new format returned by tet create function.